Extracted ChatEndedSection into its own component from SendMessages.tsx
