+++ src/components/pages/StudentsPage/SendMessages.tsx
@@ -1,11 +1,9 @@
 /** @jsxImportSource @emotion/react */
 
-import { Box, Fab, Icon, Typography } from '@mui/material';
-import { Dispatch, SetStateAction, useState, useEffect, useRef } from 'react';
+import { Box, Fab, Icon } from '@mui/material';
+import { Dispatch, SetStateAction, useRef, useState } from 'react';
 import { Socket } from 'socket.io-client';
 
-import Link from '@components/shared/Link';
-import { useStudentInClassroom } from '@hooks/useStudentInClassroom';
 import sendMessagesCSS from './SendMessages.css';
 import { PAIRED } from '@utils/classrooms';
 import { StudentPairedChat, StudentSoloChat } from './index';
@@ -14,24 +12,17 @@ interface SendMessagesProps {
   socket: Socket;
   chat: StudentPairedChat | StudentSoloChat;
   setChat: Dispatch<SetStateAction<StudentPairedChat | StudentSoloChat>>;
-  chatEndedMsg: null | string;
   setPeerIsTyping: Dispatch<SetStateAction<boolean>>;
-  classroomName: string;
-  socketId: string;
 }
 
 export default function SendMessages({
   socket,
   chat,
   setChat,
-  chatEndedMsg,
   setPeerIsTyping,
-  classroomName,
-  socketId,
 }: SendMessagesProps) {
   const typeMessageInput = useRef(null);
   const [message, setMessage] = useState('');
-  const isConnected = useStudentInClassroom(classroomName, socketId);
 
   function sendMessage(e) {
     e.preventDefault();
@@ -83,50 +74,30 @@ export default function SendMessages({
     socket.emit('student typing');
   }
 
-  let chatEndedInformationalMessage = null;
-  if (!isConnected) {
-    chatEndedInformationalMessage = (
-      <>
-        You were logged out. Return to the{' '}
-        <Link href='/'>Frempco homepage</Link> and login again.
-      </>
-    );
-  } else if (chatEndedMsg) {
-    chatEndedInformationalMessage = chatEndedMsg;
-  }
-
   return (
     <Box>
-      {!chatEndedInformationalMessage && (
-        <form onSubmit={sendMessage}>
-          <Box sx={{ textAlign: 'center' }}>
-            <input
-              css={sendMessagesCSS.message}
-              value={message}
-              placeholder={`Talk as ${chat.characters.you}...`}
-              maxLength={chat.mode === PAIRED ? 75 : 120}
-              onChange={sendUserIsTyping}
-              autoFocus
-              ref={typeMessageInput}
-            />
+      <form onSubmit={sendMessage}>
+        <Box sx={{ textAlign: 'center' }}>
+          <input
+            css={sendMessagesCSS.message}
+            value={message}
+            placeholder={`Talk as ${chat.characters.you}...`}
+            maxLength={chat.mode === PAIRED ? 75 : 120}
+            onChange={sendUserIsTyping}
+            autoFocus
+            ref={typeMessageInput}
+          />
 
-            <Fab
-              size='small'
-              type='submit'
-              color='primary'
-              style={{ marginLeft: '10px' }}
-            >
-              <Icon sx={{ fontSize: 24 }}>send</Icon>
-            </Fab>
-          </Box>
-        </form>
-      )}
-
-      {chatEndedInformationalMessage && (
-        <Typography css={sendMessagesCSS.chatEndedInfo}>
-          {chatEndedInformationalMessage}
-        </Typography>
-      )}
+          <Fab
+            size='small'
+            type='submit'
+            color='primary'
+            style={{ marginLeft: '10px' }}
+          >
+            <Icon sx={{ fontSize: 24 }}>send</Icon>
+          </Fab>
+        </Box>
+      </form>
     </Box>
   );
 }
